+++ gambit/cli/__init__.py
@@ -3,4 +3,5 @@
 from .root import cli
 from . import query
 from . import signatures
+from . import dist
 from . import debug